requirements.txt
# Core
chainlit==2.10.0
langchain==0.3.20
langchain-anthropic==0.3.12

# Google API
google-api-python-client==2.166.0
google-auth==2.38.0

# Excel processing
openpyxl==3.1.5

# Utilities
python-dotenv==1.1.0
